feat: advanced-worth-opponent-calculation
the number of games has been increased to 20, to significantly reduce chance based results.

Additionally, instead of making the untrained agent verse a fixed amount random players, and a fixed amount of skilled players, instead the number of skilled players will increase as the number of generations increases.

As per the current setup, initially the agent will play 20 random players, and each 100 generations, one of the random players will be replaced with a skilled player.

This will allow the agent to be eased into more challenging gameplay, also allowing it to achieve a higher maximum fitness.